Camera Implementation and Minor Changes
Slight changes to movement code and some organisation, and the (obvious) implementation of the first generation camera. This does come packaged with an fps lock for stability, I promise I'll try find a better way of doing this later if I have the time; but for now it works and is perfectly acceptable.